Replace the map merger with a new map converter.
