+++ tools/mapconvert/convert.bat
@@ -1,0 +1,4 @@
+@echo off
+set MAPROOT="../../_maps/"
+python mapconvert_ui.py %MAPROOT%
+pause
