--- tools/mapmerge/Run Map Merge - DMM.bat
@@ -1,5 +1,0 @@
-@echo off
-set MAPROOT="../../_maps/"
-set TGM=0
-python mapmerger.py %1 %MAPROOT% %TGM%
-pause
