--- tools/mapmerge/Run Map Merge - TGM.bat
@@ -1,5 +1,0 @@
-@echo off
-set MAPROOT="../../_maps/"
-set TGM=1
-python mapmerger.py %1 %MAPROOT% %TGM%
-pause
