--- tools/mapmerge/old_java_mapmerge/Run Map Merge.bat
@@ -1,4 +1,0 @@
-@echo off
-call java -jar MapMerge.jar "../../_maps/" /wait
-
-pause
